+++ pkg/gsutil/client.go
@@ -1,0 +1,27 @@
+package gsutil
+
+import (
+	"context"
+
+	v1 "github.com/solo-io/glooshot/pkg/api/v1"
+	"github.com/solo-io/go-utils/kubeutils"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients/factory"
+	"github.com/solo-io/solo-kit/pkg/api/v1/clients/kube"
+)
+
+func GetExperimentClient(ctx context.Context, skipCrdCreation bool) (v1.ExperimentClient, error) {
+	cfg, err := kubeutils.GetConfig("", "")
+	if err != nil {
+		return nil, err
+	}
+	cache := kube.NewKubeCache(ctx)
+	rcFactory := &factory.KubeResourceClientFactory{
+		Crd:             v1.ExperimentCrd,
+		Cfg:             cfg,
+		SharedCache:     cache,
+		SkipCrdCreation: skipCrdCreation,
+	}
+	client, err := v1.NewExperimentClient(rcFactory)
+	client.Register()
+	return client, nil
+}
