+++ pkg/version/version.go
@@ -3,4 +3,4 @@ package version
 var Version = "dev" // this value will be overwritten by the linker
 
 const AppName = "glooshot"
-const CliAppName = "glooshotctl"
+const CliAppName = "glooshot"
